+++ tool_result_selection/__init__.py
@@ -1,4 +1,0 @@
-from threedi_results_analysis.tool_result_selection.result_selection import ThreeDiResultSelection
-
-
-ThreeDiResultSelection  # flake8